sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taco"
version = "0.2.0"
description = "TACO — The A2A Construction Open-standard SDK"
readme = "../README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11",
    "a2a-sdk>=0.3.20",
]

[project.optional-dependencies]
server = ["a2a-sdk[http-server]>=0.3.20"]
client = ["httpx>=0.27"]
test = [
    "pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27",
    "a2a-sdk[http-server]>=0.3.20",
]
all = ["a2a-sdk[http-server]>=0.3.20", "httpx>=0.27"]
dev = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.27",
    "pytest>=8.0", "pytest-asyncio>=0.24",
]

[project.urls]
Homepage = "https://github.com/pelles-ai/taco"
Documentation = "https://taco-protocol.dev"
Repository = "https://github.com/pelles-ai/taco"

[tool.pytest.ini_options]
asyncio_mode = "auto"
